correcting mv errors

+++ installers/ros_ws.sh
@@ -6,6 +6,7 @@ source ../lib/extra_tools.sh
 ## Initialize variables
 INSTALL_GITMAN="true"
 BUILD="false"
+FORCE_UPDATE="false"
 # Name of workspace
 # Attention: if you change this name, change also into mavlink&mavros.sh
 NAME=skyrats_ws2
@@ -21,6 +22,9 @@ do
 	fi
 	if [[ $arg == "--build" ]]; then
 		BUILD="true"
+	fi
+	if [[ $arg == "--build" ]]; then
+		FORCE_UPDATE="true"
 	fi
 done
 
@@ -38,8 +42,29 @@ mkdir -p ~/${NAME}/src
 ## Install first test packages
 gitman install --force px4_ros_com px4_msgs
 echo "$0: moving packages to ~/${NAME}"
-sudo mv --update "$MY_PATH/../src/px4_ros_com" ~/${NAME}/src
-sudo mv --update "$MY_PATH/../src/px4_msgs" ~/${NAME}/src
+# px4_ros_com
+if [[ `find ~/${NAME}/src/px4_ros_com` != "" ]]
+then
+    if [[ FORCE_UPDATE == "true" ]]
+    then
+        sudo rm -rf ~/${NAME}/src/px4_ros_com
+        sudo mv --update "$MY_PATH/../src/px4_ros_com" ~/${NAME}/src
+    fi
+else
+    sudo mv --update "$MY_PATH/../src/px4_ros_com" ~/${NAME}/src
+fi
+# px4_msgs
+if [[ `find ~/${NAME}/src/px4_msgs` != "" ]]
+then
+    if [[ FORCE_UPDATE == "true" ]]
+    then
+        sudo rm -rf ~/${NAME}/src/px4_msgs
+        sudo mv --update "$MY_PATH/../src/px4_msgs" ~/${NAME}/src
+    fi
+else
+    sudo mv --update "$MY_PATH/../src/px4_msgs" ~/${NAME}/src
+fi
+
 
 ## Add colcon_cd dependecies in .bashrc 
 addToBashrc "source /usr/share/colcon_cd/function/colcon_cd.sh"
